pyproject.toml
[tool.poetry]
name = "wither-role"
version = "0.1.0"
description = ""
authors = ["misha kalita <mishakalita1@gmail.com>"]
readme = "README.md"
packages = [{include = "wither_role"}]

[tool.poetry.dependencies]
python = ">=3.11,<=3.12"
fastapi = "^0.110.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
yarl = "^1.9.4"
toml = "^0.10.2"
uvicorn = "^0.28.0"
alembic = "^1.13.1"
black = "^24.2.0"
dependency-injector = "^4.41.0"
jinja2 = "^3.1.3"
pyjwt = "^2.8.0"
argon2-cffi = "^23.1.0"
email-validator = "^2.1.1"
itsdangerous = "^2.2.0"
fastapi-babel = "^0.0.9"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
